# open_api_docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- {}
paths:
...
  /api/predefined-questions/{predefinedQuestion}/toggle-approval:
    post:
      tags:
      - predefined-question-controller
      operationId: toggleApproval
      parameters:
      - name: predefinedQuestion
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PredefinedQuestion"
  /api/predefined-questions/{predefinedQuestion}/suggest-fine-tuning:
    post:
      tags:
      - predefined-question-controller
      operationId: suggestQuestionForFineTuning
      parameters:
      - name: predefinedQuestion
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionSuggestionCreationParams"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SuggestedQuestionForFineTuning"
  /api/predefined-questions/{note}/refine-question:
    post:
      tags:
      - predefined-question-controller
      operationId: refineQuestion
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredefinedQuestion"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PredefinedQuestion"
  /api/predefined-questions/{note}/note-questions:
    get:
      tags:
      - predefined-question-controller
      operationId: getAllQuestionByNote
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PredefinedQuestion"
    post:
      tags:
      - predefined-question-controller
      operationId: addQuestionManually
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredefinedQuestion"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PredefinedQuestion"
  /api/predefined-questions/generate-question-without-save:
    post:
      tags:
      - predefined-question-controller
      operationId: generateQuestionWithoutSave
      parameters:
      - name: note
        in: query
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PredefinedQuestion"
...
  /api/notes/{note}/semantic-search:
    post:
      tags:
      - search-controller
      operationId: semanticSearchWithin
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchTerm"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NoteSearchResult"
  /api/notes/{note}/search:
    post:
      tags:
      - search-controller
      operationId: searchForLinkTargetWithin
      parameters:
      - name: note
        in: path
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchTerm"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NoteSearchResult"
...
  /api/notes/semantic-search:
    post:
      tags:
      - search-controller
      operationId: semanticSearch
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchTerm"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NoteSearchResult"
  /api/notes/search:
    post:
      tags:
      - search-controller
      operationId: searchForLinkTarget
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchTerm"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NoteSearchResult"
...
  /api/notebook_certificate_approvals/{notebookCertificateApproval}/approve:
    post:
      tags:
      - notebook-certificate-approval-controller
      operationId: approve
      parameters:
      - name: notebookCertificateApproval
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotebookCertificateApproval"
  /api/notebook_certificate_approvals/request-approval/{notebook}:
    post:
      tags:
      - notebook-certificate-approval-controller
      operationId: requestApprovalForNotebook
      parameters:
      - name: notebook
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotebookCertificateApproval"
...
  /api/notebook_certificate_approvals/get-all-pending-request:
    get:
      tags:
      - notebook-certificate-approval-controller
      operationId: getAllPendingRequest
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotebookCertificateApproval"
  /api/notebook_certificate_approvals/for-notebook/{notebook}:
    get:
      tags:
      - notebook-certificate-approval-controller
      operationId: getApprovalForNotebook
      parameters:
      - name: notebook
        in: path
        required: true
        schema:
          type: integer
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                type: string
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/NotebookCertificateApproval"
...
components:
  schemas:
    Circle:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          maxLength: 255
          minLength: 1
      required:
      - id
      - name
...
    MCQWithAnswer:
      type: object
      properties:
        multipleChoicesQuestion:
          $ref: "#/components/schemas/MultipleChoicesQuestion"
        correctChoiceIndex:
          type: integer
          format: int32
        strictChoiceOrder:
          type: boolean
      required:
      - correctChoiceIndex
      - multipleChoicesQuestion
      - strictChoiceOrder
    MultipleChoicesQuestion:
      type: object
      properties:
        stem:
          type: string
        choices:
          type: array
          items:
            type: string
      required:
      - choices
      - stem
...
    PredefinedQuestion:
      type: object
      properties:
        id:
          type: integer
          format: int32
        multipleChoicesQuestion:
          $ref: "#/components/schemas/MultipleChoicesQuestion"
        correctAnswerIndex:
          type: integer
          format: int32
        approved:
          type: boolean
      required:
      - id
      - multipleChoicesQuestion
...
    NoteTopology:
      type: object
      properties:
        id:
          type: integer
          format: int32
        titleOrPredicate:
          type: string
        shortDetails:
          type: string
        linkType:
          type: string
          enum:
          - no link
          - related to
          - a specialization of
          - an application of
          - an instance of
          - a part of
          - tagged by
          - an attribute of
          - the opposite of
          - author of
          - using
          - an example of
          - before
          - similar to
          - confused with
        objectNoteTopology:
          $ref: "#/components/schemas/NoteTopology"
        parentOrSubjectNoteTopology:
          $ref: "#/components/schemas/NoteTopology"
        title:
          type: string
      required:
      - id
      - titleOrPredicate
    Notebook:
      type: object
      properties:
        id:
          type: integer
          format: int32
        certifiable:
          type: boolean
        notebookSettings:
          $ref: "#/components/schemas/NotebookSettings"
        creatorId:
          type: string
        title:
          type: string
        circle:
          $ref: "#/components/schemas/Circle"
        headNoteId:
          type: integer
          format: int32
        shortDetails:
          type: string
        updated_at:
          type: string
          format: date-time
      required:
      - headNoteId
      - id
      - notebookSettings
      - title
      - updated_at
    NotebookSettings:
      type: object
      properties:
        skipMemoryTrackingEntirely:
          type: boolean
        numberOfQuestionsInAssessment:
          type: integer
          format: int32
        certificateExpiry:
          type: string
          default: 1y
          example: 1y 6m
...
    QuestionSuggestionCreationParams:
      type: object
      properties:
        comment:
          type: string
        isPositiveFeedback:
          type: boolean
    SuggestedQuestionForFineTuning:
      type: object
      properties:
        id:
          type: integer
          format: int32
        comment:
          type: string
        preservedQuestion:
          $ref: "#/components/schemas/MCQWithAnswer"
        preservedNoteContent:
          type: string
        realCorrectAnswers:
          type: string
        createdAt:
          type: string
          format: date-time
        positiveFeedback:
          type: boolean
      required:
      - id
      - preservedQuestion
...
    SearchTerm:
      type: object
      properties:
        searchKey:
          type: string
        allMyNotebooksAndSubscriptions:
          type: boolean
        allMyCircles:
          type: boolean
      required:
      - searchKey
    NoteSearchResult:
      type: object
      properties:
        noteTopology:
          $ref: "#/components/schemas/NoteTopology"
        distance:
          type: number
          format: float
      required:
      - noteTopology
...
    NotebookCertificateApproval:
      type: object
      properties:
        id:
          type: integer
          format: int32
        notebook:
          $ref: "#/components/schemas/Notebook"
        lastApprovalTime:
          type: string
          format: date-time
      required:
      - id
      - notebook
...